+++ editor/src/components/canvas/canvas-strategies/strategies/shared-absolute-resize-strategy-helpers.ts
@@ -54,7 +54,7 @@ import { addOrMergeIntendedBounds } from './shared-keyboard-strategy-helpers'
 import type { InspectorStrategy } from '../../../../components/inspector/inspector-strategies/inspector-strategy'
 import { setElementsToRerenderCommand } from '../../commands/set-elements-to-rerender-command'
 import { withUnderlyingTarget } from '../../../../components/editor/store/editor-state'
-import { pushIntendedBoundsAndUpdateGroups } from '../../commands/push-intended-bounds-and-update-targets-command'
+import { pushIntendedBoundsAndUpdateGroups } from '../../commands/push-intended-bounds-and-update-groups-command'
 
 export function createResizeCommands(
   element: JSXElement,
