+++ editor/src/components/canvas/canvas-strategies/strategies/shared-move-strategies-helpers.ts
@@ -38,7 +38,7 @@ import {
   lengthPropertyToAdjust,
 } from '../../commands/adjust-css-length-command'
 import type { CanvasCommand } from '../../commands/commands'
-import { pushIntendedBoundsAndUpdateGroups } from '../../commands/push-intended-bounds-and-update-targets-command'
+import { pushIntendedBoundsAndUpdateGroups } from '../../commands/push-intended-bounds-and-update-groups-command'
 import { setCursorCommand } from '../../commands/set-cursor-command'
 import { setElementsToRerenderCommand } from '../../commands/set-elements-to-rerender-command'
 import { setSnappingGuidelines } from '../../commands/set-snapping-guidelines-command'
